+++ webApp/pom.xml
@@ -14,7 +14,7 @@
 	<name>webApp</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>22</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
