+++ webApp/src/main/java/com/radient/ftsapp/FTSApplication.java
@@ -1,13 +1,13 @@
-package com.example.webApp;
+package com.radient.ftsapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class WebAppApplication {
+public class FTSApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(WebAppApplication.class, args);
+		SpringApplication.run(FTSApplication.class, args);
 	}
 
 }
